refactor: DBTP-1798 - move version domain code out of utils (#788)
